requirements.txt
Flask==3.0.0
Flask-SQLAlchemy==3.1.1
Flask-Login==0.6.3
Faker==22.0.0
requests==2.31.0
python-dotenv==1.0.0
beautifulsoup4==4.12.3
PyQt6==6.6.1
